chore: konverter postcss-, prettier- og sentry-config til ESM
postcss.config.cjs, .prettierrc.cjs og setup/sentry-release.cjs var de
siste CJS-filene som ikke var bundet til Jest-oppsettet. Alle tre støtter
ESM i pakken sine gjeldende versjoner.
