+++ package.json
@@ -86,7 +86,7 @@
 		"verify": "pnpm lint && pnpm format:check",
 		"up": "docker-compose up -d --remove-orphans --build",
 		"down": "docker-compose down",
-		"sentry-release": "node ./setup/sentry-release.cjs",
+		"sentry-release": "node ./setup/sentry-release.js",
 		"prepare": "husky"
 	},
 	"msw": {
